pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "trns"
version = "0.1.0"
description = "Transcription and summarization tool for YouTube, Twitter/X.com videos, and local files with automatic translation and language model processing"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Misha S.", email = "862andfurther@gmai.com"}
]
Homepage = "https://github.com/kakoyvostorg/trns"
Documentation = "https://github.com/kakoyvostorg/trns/tree/main/docs"
Repository = "https://github.com/kakoyvostorg/trns"

keywords = ["transcription", "whisper", "youtube", "telegram", "translation", "language-model"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "youtube-transcript-api>=0.6.0",
    "yt-dlp>=2023.10.7",
    "faster-whisper>=0.9.0",
    "deep-translator>=1.11.4",
    "openai>=1.0.0",
    "tqdm>=4.66.0",
    "python-telegram-bot>=20.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
trns = "trns.cli.main:main"

[tool.setuptools]
packages = ["trns", "trns.bot", "trns.transcription", "trns.cli"]

[tool.setuptools.package-data]
trns = ["*.json", "*.md", "*.txt"]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

